feat!(Makefile): new BUILD_ARG for global use
This commit introduces a new environment variable `BUILD_ARG` in the
Makefile, which allows changing a sort-of 'top-level' ENV variable 
to specify the first argument passed to the `<cargo>` command when building.
This change is intended to provide
flexibility in build configurations and allow for testing new tools 
such as 'cargo build' vs. 'cargo <zigbuild>' without multi-place
modifications.